+++ agents/agents/executor/executor_test.go
@@ -14,10 +14,10 @@ import (
 	"github.com/synapsecns/sanguine/core/merkle"
 	agentsConfig "github.com/synapsecns/sanguine/ethergo/signer/config"
 	omniClient "github.com/synapsecns/sanguine/services/omnirpc/client"
-	"github.com/synapsecns/sanguine/services/scribe/backfill"
+	"github.com/synapsecns/sanguine/services/scribe/backend"
 	"github.com/synapsecns/sanguine/services/scribe/client"
 	"github.com/synapsecns/sanguine/services/scribe/config"
-	"github.com/synapsecns/sanguine/services/scribe/node"
+	"github.com/synapsecns/sanguine/services/scribe/service"
 )
 
 func (e *ExecutorSuite) TestVerifyState() {
@@ -130,23 +130,19 @@ func (e *ExecutorSuite) TestMerkleInsert() {
 		GetLogsBatchAmount: 1,
 		StoreConcurrency:   1,
 		GetLogsRange:       1,
-		ConfirmationConfig: config.ConfirmationConfig{
-			RequiredConfirmations:   1,
-			ConfirmationThreshold:   1,
-			ConfirmationRefreshRate: 1,
-		},
-		Contracts: []config.ContractConfig{contractConfig},
+		Confirmations:      0,
+		Contracts:          []config.ContractConfig{contractConfig},
 	}
 	scribeConfig := config.Config{
 		Chains: []config.ChainConfig{chainConfig},
 	}
-	simulatedClient, err := backfill.DialBackend(e.GetTestContext(), e.TestBackendOrigin.RPCAddress(), e.ScribeMetrics)
+	simulatedClient, err := backend.DialBackend(e.GetTestContext(), e.TestBackendOrigin.RPCAddress(), e.ScribeMetrics)
 	e.Nil(err)
-	clients := map[uint32][]backfill.ScribeBackend{
+	clients := map[uint32][]backend.ScribeBackend{
 		chainID: {simulatedClient, simulatedClient},
 	}
 
-	scribe, err := node.NewScribe(e.ScribeTestDB, clients, scribeConfig, e.ScribeMetrics)
+	scribe, err := service.NewScribe(e.ScribeTestDB, clients, scribeConfig, e.ScribeMetrics)
 	e.Nil(err)
 
 	scribeClient := client.NewEmbeddedScribe("sqlite", e.DBPath, e.ScribeMetrics)
@@ -458,11 +454,11 @@ func (e *ExecutorSuite) TestExecutor() {
 
 	testContractDest, _ := e.TestDeployManager.GetAgentsTestContract(e.GetTestContext(), e.TestBackendDestination)
 
-	originClient, err := backfill.DialBackend(e.GetTestContext(), e.TestBackendOrigin.RPCAddress(), e.ScribeMetrics)
+	originClient, err := backend.DialBackend(e.GetTestContext(), e.TestBackendOrigin.RPCAddress(), e.ScribeMetrics)
 	e.Nil(err)
-	destinationClient, err := backfill.DialBackend(e.GetTestContext(), e.TestBackendDestination.RPCAddress(), e.ScribeMetrics)
+	destinationClient, err := backend.DialBackend(e.GetTestContext(), e.TestBackendDestination.RPCAddress(), e.ScribeMetrics)
 	e.Nil(err)
-	summitClient, err := backfill.DialBackend(e.GetTestContext(), e.TestBackendSummit.RPCAddress(), e.ScribeMetrics)
+	summitClient, err := backend.DialBackend(e.GetTestContext(), e.TestBackendSummit.RPCAddress(), e.ScribeMetrics)
 	e.Nil(err)
 
 	originConfig := config.ContractConfig{
@@ -474,12 +470,8 @@ func (e *ExecutorSuite) TestExecutor() {
 		GetLogsBatchAmount: 1,
 		StoreConcurrency:   1,
 		GetLogsRange:       1,
-		ConfirmationConfig: config.ConfirmationConfig{
-			RequiredConfirmations:   1,
-			ConfirmationThreshold:   1,
-			ConfirmationRefreshRate: 1,
-		},
-		Contracts: []config.ContractConfig{originConfig},
+		Confirmations:      0,
+		Contracts:          []config.ContractConfig{originConfig},
 	}
 	destinationConfig := config.ContractConfig{
 		Address:    e.DestinationContract.Address().String(),
@@ -490,11 +482,8 @@ func (e *ExecutorSuite) TestExecutor() {
 		GetLogsBatchAmount: 1,
 		StoreConcurrency:   1,
 		GetLogsRange:       1,
-		ConfirmationConfig: config.ConfirmationConfig{
-			RequiredConfirmations:   1,
-			ConfirmationThreshold:   1,
-			ConfirmationRefreshRate: 1,
-		},
+		Confirmations:      0,
+
 		Contracts: []config.ContractConfig{destinationConfig},
 	}
 	summitConfig := config.ContractConfig{
@@ -506,23 +495,20 @@ func (e *ExecutorSuite) TestExecutor() {
 		GetLogsBatchAmount: 1,
 		StoreConcurrency:   1,
 		GetLogsRange:       1,
-		ConfirmationConfig: config.ConfirmationConfig{
-			RequiredConfirmations:   1,
-			ConfirmationThreshold:   1,
-			ConfirmationRefreshRate: 1,
-		},
+		Confirmations:      0,
+
 		Contracts: []config.ContractConfig{summitConfig},
 	}
 	scribeConfig := config.Config{
 		Chains: []config.ChainConfig{originChainConfig, destinationChainConfig, summitChainConfig},
 	}
-	clients := map[uint32][]backfill.ScribeBackend{
+	clients := map[uint32][]backend.ScribeBackend{
 		chainID:     {originClient, originClient},
 		destination: {destinationClient, destinationClient},
 		summit:      {summitClient, summitClient},
 	}
 
-	scribe, err := node.NewScribe(e.ScribeTestDB, clients, scribeConfig, e.ScribeMetrics)
+	scribe, err := service.NewScribe(e.ScribeTestDB, clients, scribeConfig, e.ScribeMetrics)
 	e.Nil(err)
 
 	scribeClient := client.NewEmbeddedScribe("sqlite", e.DBPath, e.ScribeMetrics)